requirements.txt
Flask==0.10.1
pandas
matplotlib
gmplot
Pillow
numpy
gi
playsound

